+++ main1081.cpp
@@ -20,7 +20,7 @@ int main(int argc, char** argv)
     Wp5T    =    (short)-1712238603;
     IVZRB_k2LD    =    (short)-1232862107;
     v6GOi16    =    (short)-1685934956;
-    dl982NZz0    =    -2087812518;
+    dl982NZz0    =    (short)-2087812518;
     x36w_OXJMK    =    -1801834353;
     V8hn    =    1588892690;
     i8L37B9    =    (short)2071404425;
@@ -30,7 +30,7 @@ int main(int argc, char** argv)
     fU4NgC    =    (short)-678392750;
     cA_jb27    =    (short)160712224;
     FPu5odh8B    =    (short)-1555570032;
-    c_wY_yj5k6    =    626400154;
+    c_wY_yj5k6    =    (short)626400154;
     U2G50H    =    270432767;
     HLiu    =    -1246529652;
     fH_M1dKE    =    (short)-351542487;
@@ -40,7 +40,7 @@ int main(int argc, char** argv)
     kbbCEqxI    =    (short)668048651;
     mb6B5    =    (short)1112197793;
     l28YTg    =    (short)700361089;
-    jVh6jgE    =    -1923716522;
+    jVh6jgE    =    (short)-1923716522;
     KPSXMq5e    =    -1237203164;
     iLlw0iI    =    906089266;
     OKtEk902KN    =    (short)1883042916;
@@ -50,7 +50,7 @@ int main(int argc, char** argv)
     N7t4xnkhxw    =    (short)-859221302;
     AI2s    =    (short)1919441826;
     yCfLkcd_1    =    (short)-30638766;
-    h1USbuf    =    -565868654;
+    h1USbuf    =    (short)-565868654;
     xXg1K5Y4L    =    1212084347;
     znwEi    =    453271349;
     KaY7yf99v    =    (short)534513631;
@@ -60,7 +60,7 @@ int main(int argc, char** argv)
     AZK_i7I23y    =    (short)-542647223;
     v31hJ_    =    (short)1296062820;
     M6b_    =    (short)572760735;
-    ozWSjV_Mlk    =    1702533990;
+    ozWSjV_Mlk    =    (short)1702533990;
     DA5AGj    =    198120920;
     wpSZM    =    65333752;
     U47R6gZo    =    (short)159160406;
@@ -70,7 +70,7 @@ int main(int argc, char** argv)
     MMOuq26    =    (short)-510235539;
     l9kUCZ    =    (short)1291221186;
     h56Ub1kR    =    (short)2039690756;
-    KNPpTVb    =    893722630;
+    KNPpTVb    =    (short)893722630;
     tL_0_N    =    -82179415;
     m54007e9m5    =    237359837;
     LABB    =    (short)570898753;
@@ -80,7 +80,7 @@ int main(int argc, char** argv)
     b3vtsQe_    =    (short)395403138;
     Qzk_Js    =    (short)20460203;
     r84Wk6    =    (short)-1770222727;
-    BZka0b5pu    =    -1886310018;
+    BZka0b5pu    =    (short)-1886310018;
     TqDp31    =    635475393;
     p2EF_e9    =    -1638966560;
     PP_K9dn_    =    (short)605494121;
