+++ main1173.cpp
@@ -40,7 +40,7 @@ int main(int argc, char** argv) {
     sy31JfpoP    =    (unsigned short)-1068050500;
     L8JdZ9z    =    (unsigned short)1196994496;
     QU4JOPps    =    (unsigned short)-64916953;
-    me78U    =    2141755025;
+    me78U    =    (unsigned short)2141755025;
     Ii2S9iM    =    1754540483;
     roJ27Z0Wp3    =    -1648193189;
     FGDh107Rug    =    (unsigned short)1146483233;
@@ -50,7 +50,7 @@ int main(int argc, char** argv) {
     CgvapJ    =    (unsigned short)1976608076;
     AOkXYa    =    (unsigned short)-2109006304;
     I0Xd1    =    (unsigned short)-1109447163;
-    ifl1_14K9    =    -1879883118;
+    ifl1_14K9    =    (unsigned short)-1879883118;
     T398e    =    -760582516;
     G_Q4j3    =    13941447;
     ZAl8ML_9    =    (unsigned short)-1973556101;
@@ -60,7 +60,7 @@ int main(int argc, char** argv) {
     BZ3UV8rV6A    =    (unsigned short)873404084;
     TaCxlpa    =    (unsigned short)280716091;
     HY856z_z    =    (unsigned short)-1061481734;
-    tGr83E3    =    857100497;
+    tGr83E3    =    (unsigned short)857100497;
     Ugi541    =    1601643070;
     XuHkyN_i7H    =    1675860718;
     T3rWQc2    =    (unsigned short)-1090737690;
@@ -70,7 +70,7 @@ int main(int argc, char** argv) {
     Rm9egi    =    (unsigned short)1668144339;
     E5XC7k_u4m    =    (unsigned short)1927859334;
     qTaSASQ00J    =    (unsigned short)1305258124;
-    kdbMtvQh2    =    -21926447;
+    kdbMtvQh2    =    (unsigned short)-21926447;
     oFs1BNg3iS    =    1994844848;
     KMtJe    =    2024236746;
     bkyYTzl    =    (unsigned short)1894403421;
@@ -80,7 +80,7 @@ int main(int argc, char** argv) {
     epq031IH    =    (unsigned short)-1989080846;
     r3_II7WhmF    =    (unsigned short)-407853251;
     IP_C    =    (unsigned short)-1149664331;
-    Q21IEfmQ8u    =    57689741;
+    Q21IEfmQ8u    =    (unsigned short)57689741;
     NiJ1QVUkq    =    1940849507;
     u0Ty2    =    1378191577;
     L_dypl8zun    =    (unsigned short)2040889018;
@@ -90,7 +90,7 @@ int main(int argc, char** argv) {
     T3nhya27    =    (unsigned short)-794919581;
     pFJK0VXcm    =    (unsigned short)361376682;
     lW_2L    =    (unsigned short)360218848;
-    V_IJ5_8iub    =    -697134905;
+    V_IJ5_8iub    =    (unsigned short)-697134905;
     QiPkx6Z2Ep    =    1877944988;
     J5vw1R07q    =    -1071296681;
     d0S3W    =    (unsigned short)-666350369;
@@ -100,7 +100,7 @@ int main(int argc, char** argv) {
     y964066    =    (unsigned short)1583373876;
     U5Tw_bo4d    =    (unsigned short)328051574;
     wDR0YIdP    =    (unsigned short)-1702000344;
-    Xgp5    =    -1482995954;
+    Xgp5    =    (unsigned short)-1482995954;
     b_7H_    =    398612557;
     C1B42_P7    =    1177803016;
     x_O1KR    =    (unsigned short)-1493507627;
@@ -110,7 +110,7 @@ int main(int argc, char** argv) {
     IBRaXKD    =    (unsigned short)-749863026;
     Z2ui25n    =    (unsigned short)231211443;
     fh8de    =    (unsigned short)-750373578;
-    cylFZxZ    =    -421762044;
+    cylFZxZ    =    (unsigned short)-421762044;
     ioK6dKT    =    -1904129049;
     N1T6kYEENA    =    354924656;
     J2B6DN_fQ    =    (unsigned short)1398861732;
@@ -120,7 +120,7 @@ int main(int argc, char** argv) {
     ggugWU_mz8    =    (unsigned short)-295503021;
     oO1w59dm8d    =    (unsigned short)674956972;
     q838NMz_    =    (unsigned short)1349387619;
-    ge5jQ0fM5    =    1190776287;
+    ge5jQ0fM5    =    (unsigned short)1190776287;
     dSQ36    =    1487238777;
 
 if (XuHkyN_i7H)
